requirements.txt
# OpportunityScout Dependencies

# Core
anthropic>=0.40.0
pyyaml>=6.0.1
aiohttp>=3.9.0
feedparser>=6.0.10

# Telegram
python-telegram-bot>=21.0

# Database
# (sqlite3 is built-in)

# Open Brain integration
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
